Support git clone via ssh

+++ bin/startup.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 HTTPD_USER=apache
 
-# Move the SSH key to a place the apache daemon can use it
-SSH_KEY_DEST=/usr/share/httpd/.ssh/id_rsa
-cp $SSH_KEY $SSH_KEY_DEST && chown apache $SSH_KEY_DEST && chmod 400 $SSH_KEY_DEST
-export SSH_KEY=$SSH_KEY_DEST
+# Move the SSH key to a place the apache daemon can use it, if it exists
+if [ -f "$SSH_KEY" ]; then 
+  SSH_KEY_DEST=/usr/share/httpd/.ssh/id_rsa
+  cp $SSH_KEY $SSH_KEY_DEST && chown apache $SSH_KEY_DEST && chmod 400 $SSH_KEY_DEST
+  export SSH_KEY=$SSH_KEY_DEST
+fi
 
 # copy docker env variables to a place where uvicorn/apache will pick them up
 # grep is a hack to avoid overwriting home directories
